// src/reddit/dto/reddit-post.dto.ts
export class RedditPost {
  data: {
    subreddit: string;
    title: string;
    link_flair_text: string;
    is_reddit_media_domain: boolean;
    is_video: boolean;
    score: number;
    domain: string;
    ups: number;
    num_comments: number;
    subreddit_name_prefixed: string;
    over_18: boolean;
    spoiler: boolean;
    author: string;
    stickied: boolean;
    permalink: string;
    url: string;
  };
}